+++ udata/tests/test_discussions.py
@@ -422,6 +422,45 @@ class DiscussionsTest(APITestCase):
 
         self.assertEqual(len(response.json["data"]), len(discussions))
 
+    def test_list_discussions_search(self):
+        user = self.login()
+        dataset = DatasetFactory()
+
+        discussion_a = DiscussionFactory(
+            user=user,
+            subject=dataset,
+            title="discussion a",
+            discussion=[
+                Message(posted_by=user, content="another message"),
+                Message(posted_by=user, content="a message with something"),
+            ],
+        )
+        discussion_b = DiscussionFactory(
+            user=user,
+            subject=dataset,
+            title="something in title",
+            discussion=[
+                Message(posted_by=user, content="another message"),
+                Message(posted_by=user, content="there is a problem"),
+            ],
+        )
+        DiscussionFactory(
+            user=user,
+            subject=dataset,
+            title="discussion c",
+            discussion=[
+                Message(posted_by=user, content="some text"),
+                Message(posted_by=user, content="and another"),
+            ],
+        )
+
+        response = self.get(url_for("api.discussions", q="something"))
+        self.assert200(response)
+
+        self.assertEqual(len(response.json["data"]), 2)
+        self.assertEqual(discussion_b.title, response.json["data"][0]["title"])
+        self.assertEqual(discussion_a.title, response.json["data"][1]["title"])
+
     def assertIdIn(self, json_data: dict, id_: str) -> None:
         for item in json_data:
             if item["id"] == id_: